oncall_table_api/requirements.txt
Django==3.2.18
mysqlclient==2.1.1
python-dateutil
